+++ netket/experimental/dynamics/_rk_solver_structures.py
@@ -20,11 +20,24 @@ import jax
 import jax.numpy as jnp
 
 import netket as nk
+from netket import config
 from netket.utils.mpi.primitives import mpi_all_jax
 from netket.utils.struct import dataclass, field
 from netket.utils.types import Array, PyTree
 
 from . import _rk_tableau as rkt
+
+
+def maybe_jax_jit(fun, *args, **kwargs):
+    """
+    Only jit if `config.netket_experimental_disable_ode_jit` is False.
+
+    This is used to disable jitting when this config is set.
+    """
+    if config.netket_experimental_disable_ode_jit:
+        return fun
+    else:
+        return jax.jit(*args, **kwargs)
 
 
 class SolverFlags(IntFlag):
@@ -156,7 +169,7 @@ def propose_time_step(
     )
 
 
-@partial(jax.jit, static_argnames=["f", "norm_fn", "dt_limits"])
+@partial(maybe_jax_jit, static_argnames=["f", "norm_fn", "dt_limits"])
 def general_time_step_adaptive(
     tableau: rkt.TableauRKExplicit,
     f: Callable,
@@ -247,7 +260,7 @@ def general_time_step_adaptive(
     )
 
 
-@partial(jax.jit, static_argnames=["f"])
+@partial(maybe_jax_jit, static_argnames=["f"])
 def general_time_step_fixed(
     tableau: rkt.TableauRKExplicit,
     f: Callable,