+++ netket/utils/config_flags.py
@@ -203,6 +203,20 @@ config.define(
 )
 
 config.define(
+    "NETKET_EXPERIMENTAL_DISABLE_ODE_JIT",
+    bool,
+    default=False,
+    help=dedent(
+        """
+        Disables the jitting of the whole ode solver, mainly used within TDVP solvers.
+        The jitting is sometimes incompatible with GPU-based calculations, and on large
+        calculations it gives negligible speedups so it might be beneficial to disable it.
+        """
+    ),
+    runtime=True,
+)
+
+config.define(
     "NETKET_SPHINX_BUILD",
     bool,
     default=False,